requirements.txt
# RAG Pipeline - Essential Dependencies Only

# Web UI
streamlit>=1.28.0

# Image Processing
Pillow>=10.0.0

# LangChain Core Components
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# Vector Database
chromadb>=0.4.22

# PDF Processing
pypdf>=4.0.1

# Word Document Processing (.docx, .doc)
docx2txt>=0.8

# Local Multilingual Embeddings
sentence-transformers>=2.2.2

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0
